Improve Gemini support to stream responses into editor
The original PR didn't use the right API, choosing the generateContent
endpoint instead of streamGenerateContent.

<https://ai.google.dev/api/generate-content#method:-models.streamgeneratecontent>

This forced them to use the Plenary job on_exit callback to handle the
full json response.

Also added a rudimentary status window indicator to make it clear that
we are still waiting for a response. This was necessary for some of the
more powerful models which take a while to start sending data.